docs(kuzu): Add Python API docs and plan feature implementation
Adds documentation for the Kuzu Python API to serve as a reference for developing the kuzuR package.

A new file, kuzu_python_api.md, is created with details on the AsyncConnection and QueryResult classes. The active context has been updated with a new section outlining specific functions from the Python API to be reviewed and implemented in the R package.